LayoutTests/editing/spelling/spelling-contenteditable-expected.txt

This tests we can write text tests for spellcheckers. To test manually, type 'zz ' in the contenteditable <div> element and see they are marked as misspelled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


zz 
PASS 'zz' is marked as misspelled.
PASS successfullyParsed is true

TEST COMPLETE
